+++ Changelog.md
@@ -6,6 +6,10 @@
 
 ### 0.5.0 @ 06/17/2026
 
+#### :mega: New
+
+1. Provide two built-in variaints of the BBCode renderer.
+
 #### :floppy_disk: Change
 
 1. Split the BBCode renderer into a sub-package. This change does not break the previous APIs.